requirements.txt
neo4j
python-dotenv
google-generativeai
